Give exporter and builder their own files
